requirements.txt
# Before this Run: pip3 install --upgrade pip setuptools wheel
# Install with: pip3 install -r requirements.txt

setuptools
wheel
requests
openai
python-dotenv
tabulate
pandas
numpy
scikit-learn